add useFileSize composable
